Выносит моки со списком городов в отд. файл Переносит константы с макс. количеством предложений и отзывов в моки
